pyproject.toml
[tool.poetry]
name = "dbsql-labeling-app-example"
version = "0.1.0"
description = "Data Labeling app example with Python, Dash and Databricks SQL"
authors = ["Ivan Trusov"]
readme = "README.md"
packages = [{include = "dbsql_labeling_app_example"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
python-dotenv = "^1.0.0"
dash = "^2.8.1"
databricks-sql-connector = "^2.4.0"
pandas = "^1.5.3"
scikit-learn = "^1.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
isort = "^5.12.0"
faker = "^17.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"